+++ src/routes/v1/tasks/index.ts
@@ -23,8 +23,8 @@ router.patch('/:name', async (ctx: Context) => {
 router.get('/:name/logs', async (ctx: Context) => {
   const Job: MJob = ctx.models.Job;
   const where = { name: ctx.params.name }
-  const items: IJob[] = await Job.getList({ where });
-  ctx.success({ items })
+  const result: { items: IJob[] } = await Job.getList({ where });
+  ctx.success(result)
 })
 
 router.post('/:name/operate/:operate', async (ctx: Context) => {