+++ src/types/model.ts
@@ -8,6 +8,7 @@ export type Hql = {
   order?: object;
   attrs?: object;
   lean?: boolean;
+  count?: boolean;
   data?: object;
   options?: object;
   page?: number;
@@ -19,12 +20,12 @@ export interface BaseModel<T> {
    * 获取所有数据
    * @param hql 条件对象
    */
-  getAll(hql?: Hql): Promise<T[]>;
+  getAll(hql?: Hql): Promise<{ items: T[] }>;
   /**
    * 获取分页列表数据
    * @param hql 条件对象
    */
-  getList(hql?: Hql): Promise<T[]>;
+  getList(hql?: Hql): Promise<{ items: T[], total?: number, ended?: boolean }>;
   /**
    * 按条件获取一个数据详情
    * @param hql 条件对象